Player/MarioRun: getSlideStickMult 98.5->99.9% ((u8)isForceSlip() narrowing cast)
Target tests the BOOL isForceSlip() result with clrlwi. r0,r3,24 (u8
narrowing) not cmpwi; the cast matches the existing (u8)isForceSlip()
convention in MarioMove.cpp. Fixed the test AND collapsed the frame
(the -0x48 vs -0x28 inflation was a consequence of the wide-int test
path, not an independent phantom inline).

+++ src/Player/MarioRun.cpp
@@ -345,7 +345,7 @@ f32 TMario::getChangeAngleSpeed()
 // getSlideStickMult - 0x8013B8E8
 f32 TMario::getSlideStickMult()
 {
-	if (isForceSlip()) {
+	if ((u8)isForceSlip()) {
 		return mSlipParamsAll.mStickSlideMult.get();
 	}
